# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ChangeProof AI API",
    "description": "API de traçabilité et documentation automatisée par IA. SaaS B2B multi-tenant.",
    "version": "0.1.0",
    "contact": {
      "name": "ChangeProof AI",
      "url": "https://changeproof.ai"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API locale (Next.js App Router)"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Recherche globale",
        "description": "Recherche dans les titres et descriptions des changements, documents et tickets.",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 500 }, "description": "Terme de recherche" }
        ],
        "responses": {
          "200": {
            "description": "Résultats de recherche",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": { "$ref": "#/components/schemas/SearchResult" }
                  }
                }
              }
            }
          },
          "400": { "description": "Paramètre 'q' requis" },
          "401": { "description": "Non autorisé" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "Cookie de session Next-Auth"
      }
    },
    "schemas": {
...
      "SearchResult": {
        "type": "object",
        "properties": {
          "summary": { "type": "string" },
          "sources": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": { "type": "string" },
                "id": { "type": "string" },
                "title": { "type": "string" }
              }
            }
          },
          "tickets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": { "id": { "type": "string" }, "title": { "type": "string" } }
            }
          },
          "commits": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": { "id": { "type": "string" }, "message": { "type": "string" } }
            }
          },
          "documents": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": { "id": { "type": "string" }, "title": { "type": "string" } }
            }
          },
          "confidence": { "type": "number" }
        }
      }
    }
  }
}